# configs/swagger.yaml
openapi: 3.0.0
info:
  title: NestJS - Open API 3.0.0
  version: 2.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: login user with credentials
      description: Login with credentials
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auth'
        required: true
      responses:
        200:
          description: Successful login
        401:
          description: Access token is missing or invalid
        400:
          description: Invalid Credentials
        500:
          description: Interval Server
  /auth/logout:
    post:
      tags:
        - Auth
      summary: logout user
      description: Logout user with credentials
      responses:
        200:
          description: Successful Register
        401:
          description: Access token is missing or invalid
        400:
          description: Invalid Credentials
        500:
          description: Interval Server
      security:
        - Bearer: []
  /auth/refreshToken:
    post:
      tags:
        - Auth
      summary: get new access token
      description: refresh new token
      responses:
        200:
          description: Successful refresh
        401:
          description: Access token is missing or invalid
        400:
          description: Invalid Credentials
        500:
          description: Interval Server
      security:
        - Bearer: []
...
  /commons:
    get:
      tags:
        - Home
      summary: read commons
      description: read info Homepage.
      responses:
        200:
          description: Get all Sharing success
        400:
          description: Invalid status value
...
  /categories:
    get:
      tags:
        - Home
      summary: read categories
      description: read info Homepage.
      parameters:
        - name: sort
          in: query
          description: ID of Share to return
          schema:
            type: string
      responses:
        200:
          description: Get all categories success
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
  /categories/{id}:
    get:
      tags:
        - Home
      summary: read categories
      description: Categories.
      parameters:
        - name: id
          in: path
          description: id of categories to return
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get all Sharing match this categories
        400:
          description: Invalid status value
  /categories/{categories_id}/news:
    get:
      tags:
        - Home
      summary: sort sharing list by category id
      description: Categories.
      parameters:
        - name: categories_id
          in: path
          description: ID of news
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        200:
          description: Get all Sharing match this categories
        400:
          description: Invalid status value
  /courses:
    get:
      tags:
        - Home
      summary: trainingPage
      description: Get all data of TrainingPage.
      parameters:
        - name: page
          in: query
          description: ID of Share to return
          schema:
            type: string
        - name: limit
          in: query
          description: ID of Share to return
          schema:
            type: string
        - name: sort
          in: query
          description: ID of Share to return
          schema:
            type: string
        - name: search
          in: query
          description: ID of Share to return
          schema:
            type: string
      responses:
        200:
          description: Get all Training success
        400:
          description: Invalid status value
  /courses/{id}:
    get:
      tags:
        - Home
      summary: read course
      description: Returns a single Course
      parameters:
        - name: id
          in: path
          description: ID of Course to return
          required: true
          schema:
            type: string
      responses:
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
...
  /admin/categories:
    post:
      tags:
        - Categories
      summary: Create new categories
      description: This can only be done by the logged in user.
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Categories'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categories'
            application/xml:
              schema:
                $ref: '#/components/schemas/Categories'
      security:
        - Bearer: []
  /admin/categories/{id}:
    get:
      tags:
        - Categories
      summary: read categories
      description: Returns a single categories
      parameters:
        - name: id
          in: path
          description: ID of categories to return
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get data success
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    put:
      tags:
        - Categories
      summary: Update categories
      description: This can only be done by the logged in user.
      parameters:
        - name: id
          in: path
          description: id that need to be update
          required: true
          schema:
            type: string
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Categories'
      responses:
        default:
          description: successful operation
      security:
        - Bearer: []
    delete:
      tags:
        - Categories
      summary: Deletes a Categories
      description: delete a Categories
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get Categories success
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
  /admin/commons:
    get:
      tags:
        - Common
      summary: Get all data of ('training','product','mission','vision','introduce') collection
      description: This can only be done by the logged in user.
      parameters:
        - name: destination
          in: query
          description: destination is ('training','product','mission','vision','introduce') to return
          required: true
          schema:
            type: string
      responses:
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error

    post:
      tags:
        - Common
      summary: Create a new common
      description: This can on be done by the logged in user
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  enum:
                    - mission
                    - vision
                    - training
                    - produce
                    - introduce
                  default: mission
                description:
                  type: string
                  example: <p>Sgod là công ty nghiên cứu và cung cấp dịch vụ/giải pháp công nghệ ứng dụng BlockChain (chuỗi khối) và AI (trí tuệ nhân tạo) vào quản lý nhằm tăng cường khả năng bảo mật Sgod  là công ty nghiên cứu và cung cấp dịch vụ/giải pháp công nghệ ứng dụng BlockChain (chuỗi khối) và AI (trí tuệ nhân tạo) vào quản lý nhằm tăng cường khả năng bảo mật Sgod là công ty nghiên cứu và cung cấp dịch vụ/giải pháp công nghệ ứng dụng BlockChain (chuỗi khối) và AI (trí tuệ nhân tạo) vào quản lý nhằm tăng cường khả năng bảo mật.</p>
                image:
                  type: string
                  format: binary
      responses:
        200:
          description: Get data success
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
...
  /admin/courses:
    get:
      tags:
        - Course
      summary: get courses
      description: This can only be done by the logged in user.
      responses:
        200:
          description: Get all Course success
        400:
          description: Invalid status value
      security:
        - Bearer: []
    post:
      tags:
        - Course
      summary: create course
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Course'
      responses:
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
  /admin/courses/{id}:
    get:
      tags:
        - Course
      summary: Get Info course
      description: Returns a single Course
      parameters:
        - name: id
          in: path
          description: ID or Slug of Course to return
          required: true
          schema:
            type: string
      responses:
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    put:
      tags:
        - Course
      summary: Update Course
      description: This can only be done by the logged in user.
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Course'
      responses:
        default:
          description: successful operation
      security:
        - Bearer: []
    delete:
      tags:
        - Course
      summary: Deletes a Course
      description: delete a Share
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
...
components:
  schemas:
...
    Auth:
      type: object
      properties:
        username:
          type: string
          example: sgod888@gmail.com
        password:
          type: string
          example: Sgod123@
...
    Categories:
      type: object
      properties:
        name:
          type: string
          default: IOT
        parent:
          type: array
          default: ['633f85e83a5cac6e4463c804']
    Course:
      type: object
      properties:
        name:
          type: string
          example: Blockchain
        description:
          type: string
          example: Blockchain & AI
        level:
          type: string
          example: Basic
        price:
          type: string
          example: 3000000
        timeOpening:
          type: string
          example: 2022-01-01
          format: date-time
          pattern: 'DD-MM-YYYY'
        timeExpire:
          type: string
          example: 2022-12-12
          format: date-time
          pattern: 'DD-MM-YYYY'
        maxQuantity:
          type: number
          example: 30
        reducePrice:
          type: number
          example: 2
        teacher:
          type: string
          example: Sgod
        image:
          type: string
          format: binary
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer